Consolidate serial_in and serial_out in serial_node.py. Delete temporary testing files
