--- igvc_ws/src/igvc_serial/src/manual_demo.py
@@ -1,35 +1,0 @@
-#!/usr/bin/env python
-
-## Lane finder reads in the image topic and pushes out to the lane_deviation topic the current deviation from lane center.
-
-import rospy
-import json
-from igvc_msgs.msg import motors
-from sensor_msgs.msg import Joy
-
-
-manual_control_pub = rospy.Publisher("/igvc/motors_raw", motors, queue_size=10)
-
-def manual_control_callback(data):
-    #http://wiki.ros.org/joy#Microsoft_Xbox_360_Wireless_Controller_for_Linux
-
-    axes = data.axes
-
-    drivetrain_msg = motors()
-    drivetrain_msg.left = axes[1]
-    drivetrain_msg.right = axes[4]
-
-    # rospy.loginfo("manual_control callback.")
-    manual_control_pub.publish(drivetrain_msg)
-
-def manual_control_node():
-    rospy.init_node('manual_control_node', anonymous=True)
-
-    rospy.Subscriber("/joy", Joy, manual_control_callback)
-    rospy.spin()
-
-if __name__ == '__main__':
-    try:
-        manual_control_node()
-    except rospy.ROSInterruptException:
-        pass
