--- igvc_ws/src/joy/test/saved/Joy.saved
@@ -1,5 +1,0 @@
-[joy/Joy]:
-float32[] axes
-int32[] buttons
-
-
